Agregando migraciones y datos a la base de datos

+++ sistema-restaurant-Back/database/migrations/2024_03_15_154327_create_restaurantes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id(); // Columna autoincrementable para la clave primaria
             $table->unsignedBigInteger('menu_id')->nullable();
             $table->string('nombre', 100);
-            $table->integer('nit');
+            $table->bigInteger('nit');
             $table->string('direccion', 100);
             $table->integer('telefono');
             $table->string('correo', 100);
